requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.6.1
PyYAML==5.3.1
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.11
